Fix the handling of supplementary characters (characters > U+FFFF).
JNI uses a modified UTF-8 encoding. For supplementary characters, an invalid UTF-8 sequence was written to the database, which resulted in interoperability problems. The solution is to avoid UTF-8 in the native code and use the UTF-16 functions of SQLite (where possible). SQLite will then convert the UTF-16 to standards-compliant, unmodified UTF-8.

This also fixes related bugs in JDBC3PreparedStatement and improves the "out of memory" handling in the native code.

Fixed Issues:
- https://bitbucket.org/xerial/sqlite-jdbc/issues/200/wrong-utf-8-decoding-of-unicode-code , same as https://github.com/xerial/sqlite-jdbc/issues/61
- https://bitbucket.org/xerial/sqlite-jdbc/issues/144/nativedbexec-throws-an-exception-without
- https://bitbucket.org/xerial/sqlite-jdbc/issues/84/bug-in-nativedbc-bind_1blob
- https://bitbucket.org/xerial/sqlite-jdbc/issues/70/setting-a-blob-in-prepstmt