Fetch SQLite sources from multiple possible locations
Since 2013 the SQLite website has the source archive in yearly subdirectories.
Instead of just fetching from the last known location, we now fetch all known
locations in the believe that only one of them actually returns data.

As of patching the sqlite.org page responds with HTTP 302 and an empty body
when hitting non-existing URLs (it redirect to /not-found.html which gives
HTTP code 200).

To make curl fail in this case we use -L (so it follows symlinks) but also
--max-redirs set to 0 to fail on the very first redirect. Additionally we
set -f for the case the sqlite.org page starts responding with proper 404
codes.

Successfully tested back to SQLite 3.8.1. Before that, the download still
works but the test suite fails.